requirements.txt
Markdown<=3.4.4
django-ace-overlay==0.8.0
#django-admin-sortable2==2.1.9
django-admin-sortable2>=1.0.4
django-filter==23.2
django-nested-admin==4.0.2
django-polymorphic==3.1.0
django-treebeard==4.7
djangorestframework>=3.14.0
lxml<5.0
